Add bin feature and move command handlers to app/commands.rs
- Add 'bin' feature flag to conditionally compile binary-specific code
- Create app/commands.rs to hold all *_command bridge functions
- Move Output trait to top-level src/output.rs for library access
- Make helper functions public where needed for command layer:
  - get_date_format, get_date_time, format_date_time in c_date
  - get_pi, get_rounding in c_pi
  - convert_input in c_chars
- Update all output.rs files to use crate::output::Output
- Remove *_command functions from individual modules
- Update test to use library function instead of command
- Add output module to main.rs
- Add proper documentation (# Errors, # Panics) to public functions

This separation allows the library to be used without binary
dependencies while keeping the CLI bridge code isolated in app/.
