+++ src/app/output/mod.rs
@@ -3,5 +3,4 @@ mod output_trait;
 /// Output handler for formatting and displaying values.
 mod outputer;
 
-pub use output_trait::Output;
 pub use outputer::Outputer;
