+++ src/c_bytes/mod.rs
@@ -6,8 +6,9 @@ pub mod output;
 pub use encoding::BytesEncoding;
 pub use output::BytesOutput;
 
-use crate::{app::AppContext, error::GivError};
+use crate::error::GivError;
 use rand::RngCore;
+
 
 /// The default number of bytes to generate.
 pub const DEFAULT_BYTE_LENGTH: usize = 32;
@@ -21,6 +22,7 @@ pub const DEFAULT_BYTE_LENGTH: usize = 32;
 /// # Returns
 ///
 /// A result containing the generated bytes or an error.
+#[allow(dead_code)]
 fn generate_bytes(length: usize) -> Result<Vec<u8>, GivError> {
     let mut bytes = vec![0u8; length];
     let mut rng = rand::rng();
@@ -28,49 +30,6 @@ fn generate_bytes(length: usize) -> Result<Vec<u8>, GivError> {
     Ok(bytes)
 }
 
-/// The 'bytes' command handler.
-///
-/// # Arguments
-///
-/// - `length` An optional length for the bytes to generate.
-/// - `encoding` An optional encoding format for the output.
-/// - `padding` Whether to use padding for base64 encoding.
-/// - `ctx` The command context.
-///
-/// # Returns
-///
-/// A result indicating success or failure.
-///
-/// # Errors
-///
-/// Returns an error if:
-/// - Raw encoding is requested with JSON output mode.
-pub fn bytes_command(
-    length: Option<usize>,
-    encoding: Option<BytesEncoding>,
-    padding: bool,
-    ctx: &mut AppContext,
-) -> Result<(), GivError> {
-    let length = length.unwrap_or(DEFAULT_BYTE_LENGTH);
-    let encoding = encoding.unwrap_or_else(BytesEncoding::default);
-
-    // Check if raw encoding is requested with JSON output mode.
-    if matches!(encoding, BytesEncoding::Raw) && ctx.output().is_json() {
-        return Err(GivError::RawBytesNotSupportedInJson);
-    }
-
-    // Generate the random bytes.
-    let bytes = generate_bytes(length)?;
-
-    // Create output with the encoded bytes.
-    let output = BytesOutput::new(&bytes, encoding, padding);
-
-    // Output the bytes.
-    ctx.output().output(&output);
-
-    // Success.
-    Ok(())
-}
 
 // Tests.
 #[cfg(test)]
