+++ src/c_chars/output.rs
@@ -1,4 +1,4 @@
-use crate::app::output::Output;
+use crate::output::Output;
 use serde::Serialize;
 
 /// A single character conversion result.
