+++ src/c_uuid/output.rs
@@ -1,4 +1,4 @@
-use crate::app::output::Output;
+use crate::output::Output;
 use serde::Serialize;
 
 /// The output from the uuid command.
